Add WIP ssh client example

+++ examples/allocation/linux/allocation.c
@@ -4,13 +4,13 @@
 #include "hc/debug.h"
 #include "hc/compiler_rt/libc.c"
 #include "hc/linux/linux.h"
-#include "hc/linux/util.c"
 #include "hc/linux/sys.c"
 #include "hc/linux/debug.c"
+#include "hc/linux/util.c"
 #include "hc/linux/heap.c"
 #include "hc/linux/helpers/_start.c"
 
-int64_t pageSize;
+int32_t pageSize;
 #define allocator_PAGE_SIZE pageSize
 #include "hc/allocator.c"
