+++ examples/openGl/gnulinux/openGl.c
@@ -7,11 +7,11 @@
 #include "hc/debug.h"
 #include "hc/compiler_rt/libc.c"
 #include "hc/linux/linux.h"
-#include "hc/linux/util.c"
 #include "hc/linux/sys.c"
 #include "hc/linux/debug.c"
+#include "hc/linux/util.c"
 #include "hc/linux/xauth.c"
-static int64_t openGl_pageSize;
+static int32_t openGl_pageSize;
 #define x11Client_PAGE_SIZE openGl_pageSize
 #include "hc/linux/x11Client.c"
 #include "hc/linux/gnulinux/libc.so.6.h"
@@ -34,7 +34,7 @@ static int64_t openGl_pageSize;
 #include "window.c"
 
 int32_t start(hc_UNUSED int32_t argc, hc_UNUSED char **argv, char **envp) {
-    openGl_pageSize = (int64_t)getauxval(AT_PAGESZ);
+    openGl_pageSize = (int32_t)getauxval(AT_PAGESZ);
 
     int32_t status = window_init(envp);
     if (status < 0) return 1;
